Feature/login1

+++ src/main/java/com/example/backend/k3s/K3sService.java
@@ -7,6 +7,7 @@ import io.kubernetes.client.openapi.models.V1Service;
 import io.kubernetes.client.openapi.models.V1ServicePort;
 import io.kubernetes.client.openapi.models.V1ServiceSpec;
 import lombok.Data;
+import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -14,7 +15,9 @@ import java.util.List;
 import java.util.Map;
 
 @Data
+@Service
 public class K3sService {
+
 
     private String serviceName;
     private Map<String,String> selector;
@@ -51,18 +54,59 @@ public class K3sService {
                     .name(String.format("%d-%s", port.getPort(), port.getProtocol().toLowerCase()))
             );
         }
-
         V1ServiceSpec spec = new V1ServiceSpec();
         spec.setSelector(selector);
         spec.setPorts(ports);
         spec.setType(type);
         service.setSpec(spec);
+//        Network network = new Network();
 
         // create k3s service
         api.createNamespacedService(namespace, service, null, null, null,null);
     }
 
     public void delete(CoreV1Api api,String namespace) throws ApiException {
-        api.deleteNamespacedService(serviceName,namespace,null,null,null,null,null,null);
+        api.listNamespacedService(namespace, null, null, null, null, null, null, null, null, null, null).getItems().forEach(service -> {
+            if (service.getMetadata().getName().equals(serviceName)) {
+                try {
+                    api.deleteNamespacedService(serviceName, namespace, null, null, null, null, null, null);
+                } catch (ApiException e) {
+                    e.printStackTrace();
+                }
+            }
+        });
+    }
+
+    public List list(CoreV1Api api,String namespace) throws ApiException {
+        List result = new ArrayList();
+        api.listNamespacedService(namespace, null, null, null, null, null, null, null, null, null, null).getItems().forEach(service -> {
+            Map<String,Object> utilMap = new HashMap<>();
+            Map<String,String> utilMap2 = new HashMap<>();
+            utilMap2.put("Name", service.getMetadata().getName());
+            utilMap2.put("Labels", String.valueOf(service.getMetadata().getNamespace()));
+            utilMap2.put("CreateTime", String.valueOf(service.getMetadata().getCreationTimestamp()));
+            utilMap2.put("ClusterIP", String.valueOf(service.getSpec().getClusterIP()));
+            utilMap2.put("Port", String.valueOf(service.getSpec().getPorts().get(0).getPort()));
+            utilMap2.put("TargetPort", String.valueOf(service.getSpec().getPorts().get(0).getTargetPort()));
+            utilMap.put(service.getMetadata().getName(), utilMap2 );
+            result.add(utilMap);
+        });
+        return result;
+    }
+
+    public Network getNetwork(String name, CoreV1Api api) {
+//        Network network = new Network();
+//        try {
+//            V1Service service = api.readNamespacedService(name, "default", null);
+//            network.setServiceName(service.getMetadata().getName());
+//            network.setClusterIp(service.getSpec().getClusterIP());
+//            network.setTargetPort(String.valueOf(service.getSpec().getPorts().get(0).getTargetPort()));
+//            network.setCreateTime(service.getMetadata().getCreationTimestamp().toLocalDateTime());
+//            network.setPort(String.valueOf(service.getSpec().getPorts().get(0).getPort()));
+//            network.setLabels(String.valueOf(service.getMetadata().getNamespace()));
+//        } catch (ApiException e) {
+//            e.printStackTrace();
+//        }
+        return null;
     }
 }
